fix: token 인증이 되지 않으면 무한 재귀되는 버그 수정
SMS-80

+++ front-server/src/store/modules/movies/comment.js
@@ -63,7 +63,7 @@ const actions = {
         actions.getCommentList(context);
       })
       .catch((err) => {
-        if (err.response.status === 401) {
+        if (err.response.status === 401 && access) {
           refresh.actions.token_refresh()
           context.dispatch('createComment', payload)
         }
@@ -87,7 +87,7 @@ const actions = {
         actions.getCommentList(context);
       })
       .catch((err) => {
-        if (err.response.status === 401) {
+        if (err.response.status === 401 && access) {
           refresh.actions.token_refresh()
           context.dispatch('deleteComment', commentId)
         }
@@ -114,7 +114,7 @@ const actions = {
         context.commit("DELETE_UPDATE_COMMENT");
       })
       .catch((err) => {
-        if (err.response.status === 401) {
+        if (err.response.status === 401 && access) {
           refresh.actions.token_refresh()
           context.dispatch('updateComment')
         }
